Added cleanup for recent events under reporting
- Attempted to add recent events for users to view under reporting. however, there seems
to be a problem here atm. the issue is that the created recent events do not serve any
significant purpose, the same details can be viewed from the property itself.

- Added cleanup to ensure that when the time comes, we will have this ready to support
new infrastructure - the recent events details.
