+++ src/features/Rent/components/MyReports/LeaseHealthAccordion.jsx
@@ -2,39 +2,35 @@ import React from "react";
 
 import dayjs from "dayjs";
 
-import { ExpandMoreRounded, InfoOutlineRounded } from "@mui/icons-material";
+import { ExpandMoreRounded } from "@mui/icons-material";
 import {
   Accordion,
   AccordionDetails,
   AccordionSummary,
   Stack,
-  Tooltip,
   Typography,
 } from "@mui/material";
 import relativeTime from "dayjs/plugin/relativeTime";
-import StatsAccordionDetailsBlock from "features/Rent/components/Reporting/StatsAccordionDetailsBlock";
+import StatsAccordionDetailsBlock from "features/Rent/components/MyReports/StatsAccordionDetailsBlock";
 import { useCalculatePropertyStatistics } from "features/Rent/hooks/useCalculatePropertyStatistics";
 
 dayjs.extend(relativeTime);
 
-const RentCollectionAccordion = ({
+const LeaseHealthAccordion = ({
   label,
   selected,
-  properties,
   dataTour,
-  existingRents,
-  existingTenants,
+  properties,
+  existingTenants = [],
 }) => {
-  const {
-    primaryTenant,
-    averageLateRentPayment,
-    averageOnTimeRentPayment,
-    outstandingBalance,
-  } = useCalculatePropertyStatistics(
+  const { primaryTenant, leaseExiprationDate } = useCalculatePropertyStatistics(
     properties,
     selected,
     existingTenants,
-    existingRents,
+  );
+
+  const selectedTenant = existingTenants?.filter(
+    (et) => et.propertyId === selected,
   );
 
   return (
@@ -42,7 +38,6 @@ const RentCollectionAccordion = ({
       data-tour={dataTour}
       elevation={0}
       key={label}
-      defaultExpanded
       sx={{
         cursor: "default",
       }}
@@ -81,40 +76,28 @@ const RentCollectionAccordion = ({
           width="100%"
         >
           <StatsAccordionDetailsBlock
-            label="On time rent"
+            label="Lease expires in"
             // diff in days since lastMoveOut happened; does not fall below 0
-            value={`${~~(averageOnTimeRentPayment * 100)}%`}
-            caption="Last 12 months"
+            value={`${Math.max(dayjs(leaseExiprationDate).diff(dayjs(), "day"), 0) || 0} days`}
+            caption={`Occupied since ${dayjs(primaryTenant?.startDate).format("MM-DD-YYYY")}`}
           />
           <StatsAccordionDetailsBlock
-            label="Avg. days late"
-            value={`${~~averageLateRentPayment} days`}
-            caption="Last 12 months"
+            label="Tenant tenure"
+            value={`${dayjs().diff(dayjs(primaryTenant?.startDate), "day")} days`}
+            caption={`Since ${dayjs(primaryTenant?.startDate).format("MM-DD-YYYY")}`}
           />
           <StatsAccordionDetailsBlock
-            label="Outstanding balance"
-            value={`$${outstandingBalance}`}
-            caption={
-              <Stack
-                direction="row"
-                spacing={1}
-                alignItems="center"
-                justifyContent="center"
-              >
-                <Typography variant="caption">Approx</Typography>
-                <Tooltip title="Outstanding balance does not take late fees or surcharges into account">
-                  <InfoOutlineRounded
-                    color="info"
-                    sx={{ height: "1rem", width: "1rem" }}
-                  />
-                </Tooltip>
-              </Stack>
-            }
+            label="Renewal Rate"
+            // rounding support with tilda
+            // Jira - #320 Parent ticket to complete this ask
+            value={`${~~((existingTenants?.length / existingTenants?.length) * 100)}%`}
+            caption={`${selectedTenant?.length} out of ${existingTenants?.length} tenants renewed`}
           />
           <StatsAccordionDetailsBlock
-            label="Cost / Rent Ratio"
+            label="Avg lease length"
+            // Jira - #320 Parent ticket to complete this ask
             value={`${primaryTenant?.term || 0}`}
-            caption="Of annual rent income"
+            caption="Across all tenants"
             applyVariant
           />
         </Stack>
@@ -123,4 +106,4 @@ const RentCollectionAccordion = ({
   );
 };
 
-export default RentCollectionAccordion;
+export default LeaseHealthAccordion;
