+++ src/features/Rent/components/MyReports/StatsAccordionDetailsBlock.jsx
