+++ src/features/Rent/components/MyReports/StatsAccordionDetailsBlock.test.jsx
